--- a/JobScheduler.Test/AllocationTests.cs
+++ b/JobScheduler.Test/AllocationTests.cs
@@ -1,6 +1,7 @@
 ﻿using JobScheduler.Test.Utils;
 using JobScheduler.Test.Utils.CustomConstraints;
 using System.Diagnostics.CodeAnalysis;
+using System.Reflection.Metadata;
 
 namespace JobScheduler.Test;
 
@@ -84,7 +85,54 @@ internal class AllocationTests : SchedulerTestFixture
         Assert.That(() => { handle1 = Scheduler.Schedule(job); }, Is.Not.AllocatingMemory());
         Assert.That(() => { handle2 = Scheduler.Schedule(job); }, Is.Not.AllocatingMemory());
         Assert.That(() => { Scheduler.Flush(); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { handle2.Complete(); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { handle1.Complete(); }, Is.Not.AllocatingMemory());
+    }
+
+    [Test]
+    public void DependentJobsDoNotAllocate()
+    {
+        var job = new SleepJob(5); // allocates
+
+        JobHandle handle1 = default;
+        JobHandle handle2 = default;
+        // we expect the first 2 jobs to allocate
+        Assert.That(() => { handle1 = Scheduler.Schedule(job); }, Is.AllocatingMemory());
+        Assert.That(() => { handle2 = Scheduler.Schedule(job, handle1); }, Is.AllocatingMemory());
+
+        // the rest of everything should not allocate
+        Assert.That(() => { Scheduler.Flush(); }, Is.Not.AllocatingMemory());
         Assert.That(() => { handle1.Complete(); }, Is.Not.AllocatingMemory());
         Assert.That(() => { handle2.Complete(); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { handle1 = Scheduler.Schedule(job); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { handle2 = Scheduler.Schedule(job, handle1); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { Scheduler.Flush(); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { handle2.Complete(); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { handle1.Complete(); }, Is.Not.AllocatingMemory());
+    }
+
+
+    [Test]
+    public void CombinedJobsDoNotAllocate()
+    {
+        var job = new SleepJob(5); // allocates
+        var handles = new JobHandle[2];
+        JobHandle combined = default;
+
+        // we expect the first 3 jobs to allocate
+        Assert.That(() => { handles[0] = Scheduler.Schedule(job); }, Is.AllocatingMemory());
+        Assert.That(() => { handles[1] = Scheduler.Schedule(job); }, Is.AllocatingMemory());
+        Assert.That(() => { combined = Scheduler.CombineDependencies(handles); }, Is.AllocatingMemory());
+
+        // the rest of everything should not allocate
+        Assert.That(() => { Scheduler.Flush(); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { combined.Complete(); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { handles[0] = Scheduler.Schedule(job); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { handles[1] = Scheduler.Schedule(job); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { combined = Scheduler.CombineDependencies(handles); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { Scheduler.Flush(); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { combined.Complete(); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { handles[0].Complete(); }, Is.Not.AllocatingMemory());
+        Assert.That(() => { handles[1].Complete(); }, Is.Not.AllocatingMemory());
     }
 }