Apologies for the giant PR, but I didn't want to pollute the repo with too many incremental changes, since this changes how the system works on a pretty fundamental level.

Let's begin...

## Public API changes

- `JobScheduler.Schedule` now takes an optional `JobHandle` dependency to set-up a single job that must `Complete()` before the job can execute
- Adds `JobScheduler.CombineDependecies(JobHandles[] handles)`, which schedules a `JobHandle` that can only complete once all the provided `handles` have individually completed.
	+ The array is allocated by the client, and can be reused, ***as long as*** it is only touched by the client once `Complete()` is called on the `JobHandle` (or one of its dependencies). Other behavior is undefined. (For example, in a game, it could be reused the next frame, if the job is completed later on the current frame.)
- Removes pool-on-complete jobs (set-and-forget API)
	+ The new system automatically returns Job resources regardless of whether they're in-progress, so you can just set-and-forget any job without calling Complete()
	+ `JobScheduler.Schedule` no longer takes a `poolOnComplete`
	+ Removes `JobHandle.Return`
- `JobScheduler.Dispose` actually quits threads now (fixes #6)
- Removes `JobScheduler.Instance` (fixes #5)
- Can no longer `JobScheduler.Schedule` or `JobScheduler.Flush` from threads other than the main thread `JobScheduler` was instantiated on.
	+ You _shouldn't_ have done that before, but technically it would do something. Now you can't.
- Removes `IJob.Schedule` in favor of `JobScheduler.Schedule`
- Changes `JobHandle.Complete` to `JobHandle.CompleteAll` for clarity

## Tests
- Adds a ton of tests
	+ All tests that inherit from `SchedulerFixture` now run for all common thread configurations (default, 1, 2, 4, 8, 16)
- Organizes the test directory a bit

Let me know if you can think of more tests to run.

## Documentation

- Updated README.md with the new API


## `JobScheduler` implementation notes
The class now uses a custom `JobInfoPool` to handle job tracking.

+ Per each produced `JobHandle`, it stores a `Dictionary` storing a struct with...
	* A unique ID
	* A list of dependencies/a single dependency
	* The `ManualResetEvent` for the job
+ When a job's thread runs...
	* It asks the `JobInfoPool` for the job's dependencies, and then runs `Complete` on each dependency. Because jobs are queued in the FIFO order they're scheduled, dependencies are guaranteed to run before dependants. At worst, there might be a few threads waiting for one thread to complete, because those threads depend on the first thread, even while there's valid work to complete later down on the queue.
		- This could be mitigated by sorting the list by dependancy counts in `Flush()` prior to dispatching the jobs to the `ConcurrentQueue`. This strategy wouldn't help if the user was frequently calling `Flush()`, however.
		- A work-stealing implementation would completely prevent this issue!
	* The thread does its work, if present.
	* The thread tells `JobInfoPool` to recycle the entry. If someone has subscribed with `Complete`, it uses a separate dictionary to avoid returning the WaitHandle too early. Otherwise, it returns the WaitHandle to the pool and clears the dictionary entry, marking the task as officially complete.
+ When Complete is called, it first checks `JobInfoPool` for the `JobHandle`'s ID.
	* If the ID is still present in the dictionary, it assumes the job is still in flight, tells `JobInfoPool` that it has subscribed to the `WaitHandle`, and calls a `ManualResetEvent.WaitOne()`.
	* If the ID isn't in the dictionary, it assumes the job has been completed already, and returns.

`JobScheduler` uses more locks than I would prefer (for `JobInfoPool`), but I don't think there's an interlocked solution here. I'm not sure how much they impact performance, here, but I've optimized and atomized them as best I can. Let me know if you have any ideas. One note is that C#'s R/W locks all allocate seemingly randomly so we have to use normal locks.


## Questions/notes

Mostly for @genaray...

- Do you know of a good way to benchmark the scheduler? I'd like to get an idea of how bad the additional job creation overhead is. That said, I don't know what a good expected value would even be.
- If you're OK with either adding external assemblies, or copying some code over, [this package](https://github.com/Wsm2110/Faster.Map) could help with the `Dictionary`
	+ It would let us control allocations much more discretely instead of `Dictionary`'s hash-miss behavior that can infrequently allocate (but still only on `Schedule`).
- I'm pretty inexperienced in parallel programming, so let me know if there are better ways of doing what I'm doing/potential oversights I missed.