@genaray 2023-10-05
Thanks first! :) 
I took a quick look. So JobInfoPool is now responsible for pooling, correct? Uses some dictionaries though.

I think even the dictionaries could be used bluntly by normal arrays. The job id is simply incrementel, correct? In this case you could just store it in a lookup array which would make the access much faster :) 

Another question, why does JobInfo exist as a struct at all? Couldn't you store the dependencies directly in the normal job-struct and pool them from there? ^^ 

But great work, that was fast :)