@LilithSilver 2023-10-05
> I took a quick look. So JobInfoPool is now responsible for pooling, correct? Uses some dictionaries though.

Yep! And yeah, not a fan of the dictionaries but they haven't caused unexpected allocation from my initial testing.

The alternative is something like this: https://github.com/Wsm2110/Faster.Map We could either add a package reference, or alternatively just copy over the class (it's MIT licensed).


> I think even the dictionaries could be used bluntly by normal arrays. The job id is simply incrementel, correct? In this case you could just store it in a lookup array which would make the access much faster :)

I couldn't think of a way to implement a normal array lookup. The issue is that, the lookup array would end up having holes in it. For example, if we assigned jobs 0, 1, 2, 3, and 1 and 2 complete first, the array is now 0, _, _, 3, and we can't reuse those two slots until 0 is gone and we can shift over everything in the array so that it looks like 3, 4, 5, 6. 

But let me know if I'm missing something. I would LOVE a lookup array-based solution, I just haven't thought of one that will only alloc for the handles currently in-progress.

> Another question, why does JobInfo exist as a struct at all? Couldn't you store the dependencies directly in the normal job-struct and pool them from there? ^^

`JobInfoPool` works on the idea that once a job has disappeared, it is assumed complete. To check dependencies we need to do a lot of "is this job complete?" checks. 

If we used the existing JobMeta, we'd have to search through both the job queue and the work queue every single time we wanted to check if a job was complete. This would cause interlock issues and if we later implement work stealing it would get crazy!