@genaray 2023-10-05
> > I think even the dictionaries could be used bluntly by normal arrays. The job id is simply incrementel, correct? In this case you could just store it in a lookup array which would make the access much faster :)
> 
> I couldn't think of a way to implement a normal array lookup. The issue is that, the lookup array would end up having holes in it. For example, if we assigned jobs 0, 1, 2, 3, and 1 and 2 complete first, the array is now 0, _, _, 3, and we can't reuse those two slots until 0 is gone and we can shift over everything in the array so that it looks like 3, 4, 5, 6.
> 
> But let me know if I'm missing something. I would LOVE a lookup array-based solution, I just haven't thought of one that will only alloc for the handles currently in-progress.


Well we could recycle the ids once their job completed. We could simply use a `ConcurrentQueue`. 
However, we would need to ensure that recycled ids only enter the queue once their whole depdency-stack completed. So we would need some sort of intermediate Queue that stores the ids until the dependency resolved. For every potential job.
`ConcurrentQueue<int>[] resolvedDependencies;  var resolvedDependencies = resolved...[jobId];`

I hope the id is understandable, but we could recycle the job-ids... first we store completed jobids till the whole tree completed and then we flush them to the pool for being reused. This way we ensure that there no deadlocks and no holes in the array (just temporary ones) ^^ 

> If we used the existing JobMeta, we'd have to search through both the job queue and the work queue every single time we wanted to check if a job was complete. This would cause interlock issues and if we later implement work stealing it would get crazy!

Hmm... really? E.g. if every JobHandle looks like this: `struct JobHandle{ Id, ResetEvent, JobHandle[] (Or int[]?) Dependencies, Completed..}` we could just iterate the tree to see which job already finished and which not. Its basically a unification of the JobHandle and the JobInfo structs. And well that one could get checked directly without accessing the infopool multiple times. But probably I'm missing something here ^^