@LilithSilver 2023-10-05
> Well we could recycle the ids once their job completed. We could simply use a ConcurrentQueue.
However, we would need to ensure that recycled ids only enter the queue once their whole depdency-stack completed. So we would need some sort of intermediate Queue that stores the ids until the dependency resolved. For every potential job.
ConcurrentQueue<int>[] resolvedDependencies;  var resolvedDependencies = resolved...[jobId];

> I hope the id is understandable, but we could recycle the job-ids... first we store completed jobids till the whole tree completed and then we flush them to the pool for being reused. This way we ensure that there no deadlocks and no holes in the array (just temporary ones) ^^

I think I understand what you mean, but how would we ensure that the user can `Complete()` any handle, even one that has already been re-pooled? 

For example... 

```
var handle1 = Scheduler.Schedule(job1);
Scheduler.Flush();
// job1 is short lived and completes immediately, so its jobId is recycled
var handle2 = Scheduler.Schedule(job2); // takes job1's ID
Scheduler.Flush();

handle1.Complete(); // will actually complete handle2 because it recycled the ID!
```

(This solution would also cause more allocation pressure, since we have to store the full dependency tree even when jobs are totally completed, but I think that's less of an issue here than the completion issue.)

> Hmm... really? E.g. if every JobHandle looks like this: struct JobHandle{ Id, ResetEvent, JobHandle[] (Or int[]?) Dependencies, Completed..} we could just iterate the tree to see which job already finished and which not. Its basically a unification of the JobHandle and the JobInfo structs. And well that one could get checked directly without accessing the infopool multiple times. But probably I'm missing something here ^^

You're right, I think, about the dependencies!

If `JobMeta` stored the dependencies instead of `JobInfo`, we could avoid a lock on `JobScheduler.cs:80`.  We would still need to grab the info for job completion and scheduling, of course, as above. But this would make it simpler and faster! I've pushed a couple commits that should improve that, and reduce a couple other cases of locking the pool as well.

However, I think it's still more effective to have a separate `JobInfo`, which stores `WaitHandle`, that exists in the pool, rather than just a `JobHandle`. The reason for this is just from a code standpoint. If we're storing `JobHandle` but also getting the handle back from the user, we can never actually trust that the user has an updated handle, because it's a struct. We can't check things like `WaitHandle` or `Completed` because we have to acquire the newest data from the pool anyways, because threads might've done work in the meantime. So why are we giving that data to the user in the first place, and making the `JobHandle` heavier to pass around for the user? Since we have to access the info pool anyways to mark a handle complete, or to check its completion status, which with the above change is all we really use the info pool for.