@LilithSilver 2023-10-06
Out of curiosity, I left some benchmarks running between `Dictionary` and `FastMap` from that one package. They both turned out roughly the same, so I don't think it matters outside of perhaps allocation-friendliness (unless we can get an array solution up and running of course).

Interestingly, however, both benchmarks turned out mostly faster than before this PR (in the non-alloc case of course; the alloc case is a standing bug). I'm not at all sure why, since I only added overhead...?

Anyways, if we get an array solution going, I'm happy to try and benchmark it.

Before this PR (see #8):

| Method              | Threads | ExcludeFirstWave | ConcurrentJobs | Waves | Mean        | Error    | StdDev   | Gen0      | Gen1      | Allocated  |
|-------------------- |-------- |----------------- |--------------- |------ |------------:|---------:|---------:|----------:|----------:|-----------:|
| **BenchmarkParallel**   | **0**       | **True**             | **32**             | **10240** |    **343.2 ms** |  **4.13 ms** |  **3.66 ms** |         **-** |         **-** |      **600 B** |
| BenchmarkSequential | 0       | True             | 32             | 10240 |  3,732.7 ms | 17.31 ms | 16.19 ms |         - |         - |      600 B |
| **BenchmarkParallel**   | **0**       | **True**             | **128**            | **10240** |  **1,344.9 ms** | **11.85 ms** | **11.09 ms** | **1000.0000** | **1000.0000** | **54921816 B** |
| BenchmarkSequential | 0       | True             | 128            | 10240 | 14,916.6 ms | 97.26 ms | 81.22 ms |         - |         - |      600 B |



This PR with Dictionary:

| Method                | Threads | ExcludeFirstWave | ConcurrentJobs | Waves | Mean        | Error    | StdDev   | Gen0      | Gen1      | Allocated  |
|---------------------- |-------- |----------------- |--------------- |------ |------------:|---------:|---------:|----------:|----------:|-----------:|
| **BenchmarkParallel**     | **0**       | **True**             | **32**             | **10240** |    **263.6 ms** |  **1.45 ms** |  **1.36 ms** |         **-** |         **-** |      **600 B** |
| BenchmarkSequential   | 0       | True             | 32             | 10240 |  3,573.0 ms |  8.12 ms |  7.20 ms |         - |         - |      600 B |
| BenchmarkDependancies | 0       | True             | 32             | 10240 |  1,258.1 ms | 20.49 ms | 19.16 ms |         - |         - |      600 B |
| **BenchmarkParallel**     | **0**       | **True**             | **128**            | **10240** |  **1,858.2 ms** | **19.16 ms** | **17.92 ms** | **1000.0000** | **1000.0000** | **54900312 B** |
| BenchmarkSequential   | 0       | True             | 128            | 10240 | 14,192.5 ms | 29.57 ms | 26.21 ms |         - |         - |      600 B |
| BenchmarkDependancies | 0       | True             | 128            | 10240 |  7,471.6 ms | 33.29 ms | 29.51 ms | 1000.0000 | 1000.0000 | 54911064 B |


This PR with FastMap:

| Method                | Threads | ExcludeFirstWave | ConcurrentJobs | Waves | Mean        | Error    | StdDev   | Gen0      | Gen1      | Allocated  |
|---------------------- |-------- |----------------- |--------------- |------ |------------:|---------:|---------:|----------:|----------:|-----------:|
| **BenchmarkParallel**     | **0**       | **True**             | **32**             | **10240** |    **254.6 ms** |  **2.65 ms** |  **2.48 ms** |         **-** |         **-** |      **600 B** |
| BenchmarkSequential   | 0       | True             | 32             | 10240 |  3,569.3 ms | 18.08 ms | 16.91 ms |         - |         - |      600 B |
| BenchmarkDependancies | 0       | True             | 32             | 10240 |  1,275.3 ms | 14.67 ms | 13.72 ms |         - |         - |      600 B |
| **BenchmarkParallel**     | **0**       | **True**             | **128**            | **10240** |  **1,751.5 ms** | **25.95 ms** | **24.27 ms** | **1000.0000** | **1000.0000** | **54927192 B** |
| BenchmarkSequential   | 0       | True             | 128            | 10240 | 14,326.0 ms | 77.32 ms | 72.33 ms |         - |         - |      600 B |
| BenchmarkDependancies | 0       | True             | 128            | 10240 |  7,307.8 ms | 21.66 ms | 20.26 ms | 1000.0000 | 1000.0000 | 54921816 B |